Add files via upload
I changed the RoboDog VoiceController Componet UML Class Diagram to better resemble the actual code of my voiceController class for task 18. I changed name of one of the attributes to be roboDogMotionModule instead of robotHardware. I also changed all the string parameters of the methods to be const string references.  I also added a new function called seeCommandList to better help me with testing the class.